Rename file to make it less specific
