+++ src/game_logic/effect_actor_components.hpp
