+++ src/game_logic/entity_factory.cpp
@@ -40,13 +40,13 @@
 #include "game_logic/collectable_components.hpp"
 #include "game_logic/damage_components.hpp"
 #include "game_logic/dynamic_geometry_components.hpp"
+#include "game_logic/effect_actor_components.hpp"
 #include "game_logic/effect_components.hpp"
 #include "game_logic/enemy_radar.hpp"
 #include "game_logic/item_container.hpp"
 #include "game_logic/interaction/elevator.hpp"
 #include "game_logic/interaction/force_field.hpp"
 #include "game_logic/trigger_components.hpp"
-#include "game_logic/wind_blown_spiders_effect.hpp"
 
 RIGEL_DISABLE_WARNINGS
 #include <boost/fusion/adapted/std_tuple.hpp>
